Security: Reduce XSS attack surface for templated HTML (#275)
Previously, a maliciously-crafted URL could cause code to be run on the client viewing an archive as part of the not found error page. This was caused by improperly escaped string templating. This was reported by @Dedal0 — thank you!

It should be noted that wabac.js already includes a CORS policy on the error page, which mitigates the effects of this, but not completely.

This PR updates how values are set in generated HTML content: rather than being set directly in the generated HTML, they are passed to `<script>` tags as JSON-encoded strings, and then set in the relevant HTML elements on the client.

These changes were tested inside of ReplayWeb.page using a couple example payloads (e.g. `https://example.com/foobar"+alert(document.cookie);//`).